refactor components to layout component
