--- src/app/resume-layout/header/header.component.html
@@ -1,15 +1,0 @@
-<div class="container">
-    <mat-toolbar>
-        <div class="logo hide-on-mobile">
-            <div class="logo-inner">
-                <a [routerLink]="['']"><img src="./assets/img/logo.png" alt="R. Schaedler III" /></a>
-            </div>
-        </div>
-        <span class="spacer"></span>
-        <button mat-stroked-button class="special-button color-white color-hover-light-grey"
-            (click)="redirect.open('https://drive.google.com/file/d/1gPX_28rEpFXoHsIUR83QADNJUTk5HcRD/view?usp=sharing')">
-            <span class="hide-on-mobile">Download</span>
-            My Resume
-        </button>
-    </mat-toolbar>
-</div>
