--- src/app/resume-layout/header/header.component.ts
@@ -1,13 +1,0 @@
-import { Component } from '@angular/core';
-import { ExternalRedirectService } from 'src/app/core/services/external-redirect.service';
-
-@Component({
-  selector: 'app-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.scss']
-})
-export class HeaderComponent {
-
-  constructor(public redirect: ExternalRedirectService) { }
-
-}
