--- src/app/resume-layout/profile/profile.component.scss
@@ -1,14 +1,0 @@
-h1 {
-    font-weight: 300 !important;
-    margin-bottom: 0;
-}
-
-.caption {
-    font-size: 1.15rem;
-    font-weight: 100 !important;
-    letter-spacing: 0;
-}
-
-.contact-info {
-    margin-bottom: 1rem;
-}
