--- src/app/resume-layout/resume-section/resume-section.component.html
@@ -1,12 +1,0 @@
-<app-section [paddTop]="paddTop" [paddBottom]="paddBottom" [class]="color">
-    <div class="row">
-        <div class="col-md-4">
-            <h5 class="title">
-                <mat-icon inline="true">{{icon}}</mat-icon> {{title}}
-            </h5>
-        </div>
-        <div class="col-md-8">
-            <ng-content></ng-content>
-        </div>
-    </div>
-</app-section>
