--- src/app/resume-layout/resume-section/resume-section.component.scss
@@ -1,3 +1,0 @@
-app-section {
-    padding: 7rem 0rem;
-}
